LIN-19 teat: categoryFilterChip 추가, 컴포넌트 분리 및 디렉토리 구조변경
